requirements.txt
apache-beam[gcp]
google-cloud-translate
